Add handler for setting up logging
Added a log handler for setting up logging: creating log file (named
Log.txt) under the temporary folder of the operating system. Uses
standard format provided by the SimpleFormatter-class. The root logger
of the Logger-system is set to output data at the information level
and higher to the file.

The LogHandler-class is a a singleton (using the enum-pattern), but
would be prettier using EJB's @Singleton-annotation.