Add business rule data structures
Added data structures for food and nutrients. Since the names of the
objects of these types must be unique, their constructors are private
and their builder-method and class are package-private to allow another
class, only in the same package, to create them. The classes implement
an empty interface, for future type info.