Add smoke testing
Added gradle smokeTest-task which depends on assemble. It runs all
JUnit-tests annotated with the @TagSmoke-tag. The value it uses,
TagValues.SMOKE_TEST is imported into the build-script for easy
modification by adding the classpath of test/ where is resides. The
smokeTest-task is run as the last step of the build-task.