Add food factory
Added factory for creating food. Creating food requires a food builder
with a unique name, otherwise an IllegalArgumentException is thrown.

Added a method for building upon an existing food object without
modifying it so that they can be updated in factory. Made it and the
normal builder method public since they will need to be accessed
outside the package.